requirements.txt
langchain_community
langchain_core
langchain_openai
sqlalchemy
mysql-connector-python
python-dotenv
streamlit
